requirements.txt
streamlit~=1.44.1
langchain-openai~=0.3.12
langchain~=0.3.22
langchain-core~=0.3.50
python-dotenv~=1.0.1